# app/api-docs/swagger.yaml
openapi: 3.0.3
info:
  title: Figma AI Ticket Generator API
  description: |
    🎯 **Unified Documentation Generation API**
    
    Generate JIRA tickets, Wiki pages, Code documentation, and more from Figma designs using AI.
    
    ## 🚀 **Key Features:**
    - **Unified Generation**: Single endpoint for all documentation formats
    - **Simplified Strategies**: AI-Powered (primary), Emergency (fallback)
    - **Format Support**: JIRA (active), Wiki/Code/Markdown (planned)
    - **Smart Auto-Detection**: Automatically selects optimal approach based on input
    - **Legacy Compatibility**: Backward compatible with deprecation warnings
    
    ## 🎯 **Primary Endpoint:**
    Use `POST /api/generate` for all documentation generation needs.
    
    ## 📋 **Migration Guide:**
    - `POST /api/generate-ticket` → `POST /api/generate` with `format: "jira"` and `strategy: "ai-powered"`
    - `POST /api/generate-ai-ticket-direct` → `POST /api/generate` with `format: "jira"` and `strategy: "ai-powered"`
    
  version: 2.0.0
  contact:
    name: Figma AI Support
    url: https://github.com/figma-ai/ticket-generator
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development Server
  - url: https://api.figma-ai.com
    description: Production Server
...
paths:
...
  /api/context/extract:
    post:
      tags: [context-intelligence]
      summary: 🧠 Extract Context from Figma Data
      description: |
        **NEW**: Extract comprehensive context intelligence from Figma data.
        
        This endpoint processes Figma URLs or frame data to extract:
        - Component hierarchy and relationships
        - Design tokens (colors, typography, spacing)
        - Layout patterns and structure
        - Accessibility insights
        - Interaction flows
        
        Uses the enhanced Context Manager with intelligent caching.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                figmaUrl:
                  type: string
                  format: uri
                  description: Figma file URL
                  example: "https://www.figma.com/file/ABC123/Design-System"
                fileKey:
                  type: string
                  description: Alternative to figmaUrl - just the file key
                  example: "ABC123"
                frameData:
                  type: array
                  description: Optional frame data for focused analysis
                  items:
                    type: object
              required: [figmaUrl]
            examples:
              basic_extraction:
                summary: Basic Context Extraction
                value:
                  figmaUrl: "https://www.figma.com/file/ABC123/Design-System"
              test_mode:
                summary: Test Mode (Mock Data)
                value:
                  fileKey: "test"
      responses:
        '200':
          description: Context extraction completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Success"
                  data:
                    type: object
                    properties:
                      context:
                        type: object
                        properties:
                          components:
                            type: array
                            items:
                              type: string
                            example: ["Button", "Input", "Card"]
                          colors:
                            type: array
                            items:
                              type: string
                            example: ["#007bff", "#6c757d", "#28a745"]
                          typography:
                            type: array
                            items:
                              type: string
                            example: ["Inter", "Roboto"]
                          layout:
                            type: string
                            example: "responsive"
                      source:
                        type: string
                        example: "context-intelligence"
                      timestamp:
                        type: string
                        format: date-time
                  metadata:
                    type: object
                    properties:
                      timestamp:
                        type: string
                        format: date-time
                      route:
                        type: string
                        example: "FigmaCore"
        '400':
          description: Missing required parameters
        '503':
          description: Context manager not available
  /api/context/get:
    get:
      tags: [context-intelligence]
      summary: 🗂️ Get Cached Context Data
      description: |
        Retrieve previously extracted and cached context data for a Figma file.
        Uses Redis caching for improved performance.
      parameters:
        - name: fileKey
          in: query
          required: true
          description: Figma file key
          schema:
            type: string
            example: "ABC123"
      responses:
        '200':
          description: Context data retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      context:
                        type: object
                        nullable: true
                      cached:
                        type: boolean
                      message:
                        type: string
                      timestamp:
                        type: string
                        format: date-time
  /api/context/search:
    get:
      tags: [context-intelligence]
      summary: 🔍 Search Context Data
      description: |
        Search through cached context data to find specific components, patterns, or design tokens.
        Supports fuzzy matching and confidence scoring.
      parameters:
        - name: q
          in: query
          required: true
          description: Search query
          schema:
            type: string
            example: "button"
      responses:
        '200':
          description: Search completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      query:
                        type: string
                      results:
                        type: array
                        items:
                          type: object
                          properties:
                            fileKey:
                              type: string
                            fileName:
                              type: string
                            match:
                              type: string
                            confidence:
                              type: number
                            contextType:
                              type: string
                      total:
                        type: integer
                      timestamp:
                        type: string
                        format: date-time
  /api/context/explore:
    get:
      tags: [context-intelligence]
      summary: 🕸️ Explore Context Relationships
      description: |
        Explore relationships and patterns within a specific context.
        Provides insights into component usage, design patterns, and recommendations.
      parameters:
        - name: fileKey
          in: query
          required: true
          description: Figma file key to explore
          schema:
            type: string
            example: "ABC123"
      responses:
        '200':
          description: Context exploration completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      fileKey:
                        type: string
                      relationships:
                        type: array
                        items:
                          type: object
                          properties:
                            type:
                              type: string
                            target:
                              type: string
                            strength:
                              type: number
                      patterns:
                        type: array
                        items:
                          type: object
                          properties:
                            pattern:
                              type: string
                            confidence:
                              type: number
                      suggestions:
                        type: array
                        items:
                          type: string
...
  /api/intelligence/semantic:
    post:
      tags: [context-intelligence]
      summary: 🧠 Semantic Analysis
      description: |
        Perform semantic analysis on design components to understand their purpose,
        meaning, and user journey context.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                figmaUrl:
                  type: string
                  format: uri
                content:
                  type: string
                  description: Alternative content to analyze
              required: [figmaUrl]
      responses:
        '200':
          description: Semantic analysis completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      semanticTags:
                        type: array
                        items:
                          type: string
                        example: ["navigation", "form", "button", "input"]
                      meaning:
                        type: string
                        example: "Login form with primary call-to-action"
                      intent:
                        type: string
                        example: "User authentication interface"
                      userJourney:
                        type: string
                        example: "Account access entry point"
                      accessibility:
                        type: object
                        properties:
                          role:
                            type: string
                          ariaLabel:
                            type: string
                          keyboardNavigable:
                            type: boolean
                      confidence:
                        type: number
                        example: 0.89
  /api/intelligence/accessibility:
    post:
      tags: [context-intelligence]
      summary: ♿ Accessibility Analysis
      description: |
        Perform comprehensive accessibility analysis including WCAG compliance,
        contrast ratios, and accessibility recommendations.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                figmaUrl:
                  type: string
                  format: uri
                content:
                  type: string
              required: [figmaUrl]
      responses:
        '200':
          description: Accessibility analysis completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      score:
                        type: integer
                        example: 85
                      issues:
                        type: array
                        items:
                          type: object
                          properties:
                            type:
                              type: string
                            severity:
                              type: string
                            message:
                              type: string
                      recommendations:
                        type: array
                        items:
                          type: string
                      wcagLevel:
                        type: string
                        example: "AA"
                      compliant:
                        type: boolean
  /api/intelligence/design-tokens:
    post:
      tags: [context-intelligence]
      summary: 🎨 Design Tokens Extraction
      description: |
        Extract and categorize design tokens including colors, typography,
        spacing, and effects from design components.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                figmaUrl:
                  type: string
                  format: uri
                content:
                  type: string
              required: [figmaUrl]
      responses:
        '200':
          description: Design tokens extracted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      colors:
                        type: array
                        items:
                          type: object
                          properties:
                            name:
                              type: string
                            value:
                              type: string
                            usage:
                              type: string
                      typography:
                        type: array
                        items:
                          type: object
                          properties:
                            name:
                              type: string
                            size:
                              type: string
                            weight:
                              type: string
                            family:
                              type: string
                      spacing:
                        type: array
                        items:
                          type: object
                          properties:
                            name:
                              type: string
                            value:
                              type: string
                      effects:
                        type: array
                        items:
                          type: object
                          properties:
                            name:
                              type: string
                            value:
                              type: string

  # 🧩 COMPONENT ANALYSIS API
  /api/component/analyze:
    post:
      tags: [context-intelligence]
      summary: 🧩 Component Structure Analysis
      description: |
        Analyze component structure, variants, states, and properties to understand
        component architecture and usage patterns.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                figmaUrl:
                  type: string
                  format: uri
                componentData:
                  type: string
                  description: Component-specific data to analyze
              required: [figmaUrl]
      responses:
        '200':
          description: Component analysis completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      type:
                        type: string
                        example: "button"
                      category:
                        type: string
                        example: "interactive"
                      atomicLevel:
                        type: string
                        example: "atom"
                      variants:
                        type: array
                        items:
                          type: string
                        example: ["primary", "secondary", "ghost"]
                      states:
                        type: array
                        items:
                          type: string
                        example: ["default", "hover", "active", "disabled"]
                      properties:
                        type: object
                        properties:
                          size:
                            type: array
                            items:
                              type: string
                          color:
                            type: array
                            items:
                              type: string
                      usage:
                        type: string
                      accessibility:
                        type: object
                        properties:
                          role:
                            type: string
                          keyboard:
                            type: boolean
                          screenReader:
                            type: boolean
  /api/component/tokens:
    post:
      tags: [context-intelligence]
      summary: 🎯 Component-Specific Tokens
      description: |
        Extract design tokens that are specific to a particular component,
        including component-level colors, spacing, and styling properties.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                figmaUrl:
                  type: string
                  format: uri
                componentData:
                  type: string
              required: [figmaUrl]
      responses:
        '200':
          description: Component tokens extracted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      component:
                        type: string
                        example: "Button"
                      tokens:
                        type: object
                        properties:
                          colors:
                            type: array
                            items:
                              type: object
                              properties:
                                property:
                                  type: string
                                token:
                                  type: string
                                value:
                                  type: string
                          typography:
                            type: array
                            items:
                              type: object
                              properties:
                                property:
                                  type: string
                                token:
                                  type: string
                                value:
                                  type: string
                          spacing:
                            type: array
                            items:
                              type: object
                              properties:
                                property:
                                  type: string
                                token:
                                  type: string
                                value:
                                  type: string
                          borders:
                            type: array
                            items:
                              type: object
                              properties:
                                property:
                                  type: string
                                token:
                                  type: string
                                value:
                                  type: string
...
components:
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (when enabled)
security:
  - {} # No security required for development
  # - ApiKeyAuth: [] # Uncomment for production
...