# app/routes/context/intelligence.js
/**
 * Context Intelligence Routes
 * Provides advanced context analysis endpoints for the unified dashboard
 */

import { BaseRoute } from '../BaseRoute.js';

export class ContextIntelligenceRoutes extends BaseRoute {
  constructor(services) {
    super(services);
    this.contextManager = services.contextManager;
    this.visualAIService = services.visualAIService;
  }

  registerRoutes(router) {
    // Context extraction and analysis
    router.post('/api/context/extract', this.handleExtractContext.bind(this));
    router.get('/api/context/get', this.handleGetContext.bind(this));
    router.get('/api/context/search', this.handleSearchContext.bind(this));
    router.get('/api/context/explore', this.handleExploreContext.bind(this));
    router.get('/api/context/summary', this.handleContextSummary.bind(this));

    // Intelligence endpoints
    router.post('/api/intelligence/semantic', this.handleSemanticAnalysis.bind(this));
    router.post('/api/intelligence/accessibility', this.handleAccessibilityAnalysis.bind(this));
    router.post('/api/intelligence/design-tokens', this.handleDesignTokens.bind(this));

    // Component analysis
    router.post('/api/component/analyze', this.handleComponentAnalysis.bind(this));
    router.post('/api/component/tokens', this.handleComponentTokens.bind(this));

    this.logger.info('✅ Context Intelligence routes registered');
  }

  /**
   * Extract context from Figma data
   * POST /api/context/extract
   */
  async handleExtractContext(req, res) {
    this.logAccess(req, 'extractContext');

    try {
      const { figmaUrl, fileKey, frameData } = req.body;

      if (!figmaUrl && !fileKey) {
        return this.sendError(res, 'figmaUrl or fileKey is required', 400);
      }

      // Prepare figma data for context extraction
      const figmaData = {
        url: figmaUrl || `https://www.figma.com/file/${fileKey}`,
        nodes: frameData || [],
        document: {
          id: fileKey || 'extracted-file',
          name: 'Extracted File',
          children: frameData || []
        },
        metadata: {
          extractedAt: new Date().toISOString(),
          source: 'context-intelligence-api'
        }
      };

      // Extract context using Context Manager
      const contextResult = await this.contextManager.extractContext(figmaData);

      this.sendSuccess(res, {
        context: contextResult,
        source: 'context-intelligence',
        timestamp: new Date().toISOString()
      });

    } catch (error) {
      this.logger.error('Context extraction failed:', error);
      this.sendError(res, `Context extraction failed: ${error.message}`, 500);
    }
  }

  /**
   * Get cached context data
   * GET /api/context/get
   */
  async handleGetContext(req, res) {
    this.logAccess(req, 'getContext');

    try {
      const { fileKey } = req.query;

      if (!fileKey) {
        return this.sendError(res, 'fileKey is required', 400);
      }

      const redis = this.getService('redis');
      const cacheKey = `figma-context-${fileKey}`;

      let contextData = await redis.get(cacheKey);

      if (!contextData) {
        return this.sendSuccess(res, {
          context: null,
          cached: false,
          message: 'No cached context found for this file'
        });
      }

      contextData = JSON.parse(contextData);

      this.sendSuccess(res, {
        context: contextData,
        cached: true,
        timestamp: new Date().toISOString()
      });

    } catch (error) {
      this.logger.error('Get context failed:', error);
      this.sendError(res, `Get context failed: ${error.message}`, 500);
    }
  }

  /**
   * Search context data
   * GET /api/context/search
   */
  async handleSearchContext(req, res) {
    this.logAccess(req, 'searchContext');

    try {
      const { q: query } = req.query;

      if (!query) {
        return this.sendError(res, 'Query parameter q is required', 400);
      }

      // For demo purposes, return mock search results
      // In production, this would search through cached context data
      const mockResults = [
        {
          fileKey: 'mock-file-1',
          fileName: 'Design System Components',
          match: 'Primary Button component found',
          confidence: 0.95,
          contextType: 'component'
        },
        {
          fileKey: 'mock-file-2',
          fileName: 'Marketing Pages',
          match: 'Color tokens detected',
          confidence: 0.87,
          contextType: 'design-tokens'
        }
      ];

      this.sendSuccess(res, {
        query: query,
        results: mockResults,
        total: mockResults.length,
        timestamp: new Date().toISOString()
      });

    } catch (error) {
      this.logger.error('Context search failed:', error);
      this.sendError(res, `Context search failed: ${error.message}`, 500);
    }
  }

  /**
   * Explore context relationships
   * GET /api/context/explore
   */
  async handleExploreContext(req, res) {
    this.logAccess(req, 'exploreContext');

    try {
      const { fileKey } = req.query;

      if (!fileKey) {
        return this.sendError(res, 'fileKey is required', 400);
      }

      // Mock exploration data
      const explorationData = {
        fileKey: fileKey,
        relationships: [
          { type: 'contains', target: 'Button Component', strength: 0.9 },
          { type: 'uses', target: 'Primary Color Token', strength: 0.8 },
          { type: 'inherits', target: 'Base Typography', strength: 0.7 }
        ],
        patterns: [
          { pattern: 'Design System Usage', confidence: 0.85 },
          { pattern: 'Atomic Design Structure', confidence: 0.72 }
        ],
        suggestions: [
          'Consider extracting common button patterns',
          'Standardize color token naming conventions'
        ]
      };

      this.sendSuccess(res, explorationData);

    } catch (error) {
      this.logger.error('Context exploration failed:', error);
      this.sendError(res, `Context exploration failed: ${error.message}`, 500);
    }
  }

  /**
   * Get context summary
   * GET /api/context/summary
   */
  async handleContextSummary(req, res) {
    this.logAccess(req, 'contextSummary');

    try {
      const redis = this.getService('redis');

      // Get summary of all cached contexts
      const summary = {
        totalFiles: 0,
        totalComponents: 0,
        totalTokens: 0,
        recentActivity: [],
        topPatterns: [
          { pattern: 'Button Components', count: 15 },
          { pattern: 'Color Tokens', count: 23 },
          { pattern: 'Typography Styles', count: 8 }
        ]
      };

      this.sendSuccess(res, summary);

    } catch (error) {
      this.logger.error('Context summary failed:', error);
      this.sendError(res, `Context summary failed: ${error.message}`, 500);
    }
  }

  /**
   * Perform semantic analysis
   * POST /api/intelligence/semantic
   */
  async handleSemanticAnalysis(req, res) {
    this.logAccess(req, 'semanticAnalysis');

    try {
      const { figmaUrl, content } = req.body;

      if (!figmaUrl && !content) {
        return this.sendError(res, 'figmaUrl or content is required', 400);
      }

      const analysis = {
        semanticTags: ['navigation', 'form', 'button', 'input'],
        meaning: 'Login form with primary call-to-action',
        intent: 'User authentication interface',
        userJourney: 'Account access entry point',
        accessibility: {
          role: 'form',
          ariaLabel: 'Login form',
          keyboardNavigable: true
        },
        confidence: 0.89
      };

      this.sendSuccess(res, analysis);

    } catch (error) {
      this.logger.error('Semantic analysis failed:', error);
      this.sendError(res, `Semantic analysis failed: ${error.message}`, 500);
    }
  }

  /**
   * Perform accessibility analysis
   * POST /api/intelligence/accessibility
   */
  async handleAccessibilityAnalysis(req, res) {
    this.logAccess(req, 'accessibilityAnalysis');

    try {
      const { figmaUrl, content } = req.body;

      const analysis = {
        score: 85,
        issues: [
          { type: 'contrast', severity: 'medium', message: 'Text contrast could be improved' },
          { type: 'alt-text', severity: 'low', message: 'Consider adding alt text for icons' }
        ],
        recommendations: [
          'Increase text contrast ratio to 4.5:1',
          'Add ARIA labels for interactive elements',
          'Ensure keyboard navigation order is logical'
        ],
        wcagLevel: 'AA',
        compliant: true
      };

      this.sendSuccess(res, analysis);

    } catch (error) {
      this.logger.error('Accessibility analysis failed:', error);
      this.sendError(res, `Accessibility analysis failed: ${error.message}`, 500);
    }
  }

  /**
   * Extract design tokens
   * POST /api/intelligence/design-tokens
   */
  async handleDesignTokens(req, res) {
    this.logAccess(req, 'designTokens');

    try {
      const { figmaUrl, content } = req.body;

      const tokens = {
        colors: [
          { name: 'primary-blue', value: '#007bff', usage: 'primary actions' },
          { name: 'secondary-gray', value: '#6c757d', usage: 'secondary elements' },
          { name: 'success-green', value: '#28a745', usage: 'success states' }
        ],
        typography: [
          { name: 'heading-large', size: '2rem', weight: 'bold', family: 'Inter' },
          { name: 'body-text', size: '1rem', weight: 'normal', family: 'Inter' }
        ],
        spacing: [
          { name: 'space-sm', value: '8px' },
          { name: 'space-md', value: '16px' },
          { name: 'space-lg', value: '24px' }
        ],
        effects: [
          { name: 'shadow-card', value: '0 2px 4px rgba(0,0,0,0.1)' }
        ]
      };

      this.sendSuccess(res, tokens);

    } catch (error) {
      this.logger.error('Design tokens extraction failed:', error);
      this.sendError(res, `Design tokens extraction failed: ${error.message}`, 500);
    }
  }

  /**
   * Analyze component structure
   * POST /api/component/analyze
   */
  async handleComponentAnalysis(req, res) {
    this.logAccess(req, 'componentAnalysis');

    try {
      const { figmaUrl, componentData } = req.body;

      const analysis = {
        type: 'button',
        category: 'interactive',
        atomicLevel: 'atom',
        variants: ['primary', 'secondary', 'ghost'],
        states: ['default', 'hover', 'active', 'disabled'],
        properties: {
          size: ['small', 'medium', 'large'],
          color: ['primary', 'secondary', 'danger']
        },
        usage: 'Call-to-action element for user interactions',
        accessibility: {
          role: 'button',
          keyboard: true,
          screenReader: true
        }
      };

      this.sendSuccess(res, analysis);

    } catch (error) {
      this.logger.error('Component analysis failed:', error);
      this.sendError(res, `Component analysis failed: ${error.message}`, 500);
    }
  }

  /**
   * Extract component-specific tokens
   * POST /api/component/tokens
   */
  async handleComponentTokens(req, res) {
    this.logAccess(req, 'componentTokens');

    try {
      const { figmaUrl, componentData } = req.body;

      const tokens = {
        component: 'Button',
        tokens: {
          colors: [
            { property: 'background-color', token: 'primary-blue', value: '#007bff' },
            { property: 'text-color', token: 'white', value: '#ffffff' }
          ],
          typography: [
            { property: 'font-size', token: 'button-text', value: '16px' },
            { property: 'font-weight', token: 'medium', value: '500' }
          ],
          spacing: [
            { property: 'padding-x', token: 'space-md', value: '16px' },
            { property: 'padding-y', token: 'space-sm', value: '8px' }
          ],
          borders: [
            { property: 'border-radius', token: 'radius-sm', value: '4px' }
          ]
        }
      };

      this.sendSuccess(res, tokens);

    } catch (error) {
      this.logger.error('Component tokens extraction failed:', error);
      this.sendError(res, `Component tokens extraction failed: ${error.message}`, 500);
    }
  }
}

export default ContextIntelligenceRoutes;